half on batch

+++ src/data/make_dataset.py
@@ -133,8 +133,8 @@ class DatasetVesuvius(Dataset):
             torch.manual_seed(seed)
             mask = torch.squeeze(self.transforms(mask))
 
-        image = image.half()
-        mask = mask.half()
+        # image = image.half()
+        # mask = mask.half()
 
         return fragment, bbox, mask, image
